Multiple test running problems were fixed. Version 0.1.1 was published

+++ src/main/java/settings/Environment.java
@@ -7,8 +7,8 @@ public class Environment {
     String appPath;
     String appPackage;
     String appActivity;
-    Integer testaiConfidenceThreshold;
-    Integer imageMatchThreshold;
+    Double testaiConfidenceThreshold;
+    Double imageMatchThreshold;
     Integer elementTimeout;
     Double platformVersion;
     JsonArray devices;
